requirements.txt
fastmcp>=2.11.0
openai>=1.14.0
mistralai>=0.1.8
tqdm>=4.66
rich>=13.7
python-dotenv>=1.0
requests>=2.31
